+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<major>`.`<minor>`.`<patch>`)
 
 ## [v2.0.1]
+### Added
+* #71 Add `pep8-naming` to linting toolchain
+
 ### Fixed
 * #70 Fix incorrect column index for missing return annotations when other annotations are present on the same line of source